Add equality tests generic over array size
Adds two equality checking methods that can be used
between ResultSets with different on-stack range
storage size.

On-stack range storage size should be an
implementation detail, and shouldn't affect equality.
But a fully generic PartialEq implementation
is a breaking change due to type inference. The
code for the PartialEq impl has been added in a
comment, in case it's ever needed.